% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{.util_optim}
\alias{.util_optim}
\title{Helper functions}
\usage{
.util_optim(
  inputvec,
  ret_age,
  ra,
  delta,
  beta,
  c_age,
  gender,
  gender_mortalityTable,
  w0,
  CF,
  li,
  lg,
  c1,
  s1,
  s2,
  s3,
  w2,
  rho2,
  rho3,
  ret,
  retr,
  psi,
  verbose = FALSE,
  warnings = FALSE
)
}
\arguments{
\item{inputvec}{c(c, c2, nu2, nu3, alpha, w3)}
}
\description{
Helper 1: Optimization function
}
