% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTILALL.R
\name{.util_optim2}
\alias{.util_optim2}
\title{Helper 1: Optimization function}
\usage{
.util_optim2(
  inputvec,
  ret_age,
  ra,
  delta,
  beta,
  c_age,
  gender,
  gender_mortalityTable2,
  w0,
  CF,
  li,
  lg,
  c1,
  s1,
  s2,
  s3,
  rho2,
  rho3,
  ret,
  retr,
  SPFretsel,
  psi,
  verbose = FALSE,
  warnings = FALSE
)
}
\arguments{
\item{inputvec}{c(c, c2, nu2, nu3, alpha, w3)}
}
\description{
Helper 1: Optimization function
}
