% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{.util_optim_w}
\alias{.util_optim_w}
\title{Helper 3: Optimize only portfolio weights}
\usage{
.util_optim_w(
  inputvec,
  ret_age,
  cc,
  c2,
  nu2,
  nu3,
  ra,
  delta,
  alpha,
  beta,
  c_age,
  gender,
  gender_mortalityTable,
  w0,
  CF,
  li,
  lg,
  c1,
  s1,
  s2,
  s3,
  w2,
  rho2,
  rho3,
  ret,
  retr,
  psi,
  verbose = FALSE,
  warnings = FALSE
)
}
\arguments{
\item{inputvec}{c(w3)}
}
\description{
Helper 3: Optimize only portfolio weights
}
