% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTILALL.R
\name{.util_optim_w2}
\alias{.util_optim_w2}
\title{Helper 3: Optimize only portfolio weights}
\usage{
.util_optim_w2(
  inputvec,
  ret_age,
  cc,
  c2,
  nu2,
  nu3,
  ra,
  delta,
  alpha,
  beta,
  c_age,
  gender,
  gender_mortalityTable2,
  w0,
  CF,
  li,
  lg,
  c1,
  s1,
  s2,
  s3,
  rho2,
  rho3,
  ret,
  retr,
  SPFretsel,
  psi,
  verbose = FALSE,
  warnings = FALSE
)
}
\arguments{
\item{inputvec}{c(w3)}
}
\description{
Helper 3: Optimize only portfolio weights
}
