% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTILALL.R
\name{.util_optim_wc2}
\alias{.util_optim_wc2}
\title{Helper 2: Optimize portfolio weights and consumption c,}
\usage{
.util_optim_wc2(
  inputvec,
  ret_age,
  c2,
  nu2,
  nu3,
  ra,
  delta,
  beta,
  c_age,
  gender,
  gender_mortalityTable2,
  w0,
  CF,
  li,
  lg,
  c1,
  s1,
  s2,
  s3,
  rho2,
  rho3,
  ret,
  retr,
  SPFretsel,
  psi,
  verbose = FALSE,
  warnings = FALSE
)
}
\arguments{
\item{inputvec}{c(c, alpha, w3)}
}
\description{
Helper 2: Optimize portfolio weights and consumption c,
}
