% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{FINDATA}
\alias{FINDATA}
\title{Various financial variables}
\format{A time series (xts) object with 588 rows and 8 variables (all described in the vignette `vignette("INPUT_returns")`)}
\source{
THOMSON REUTERS DATASTREAM
}
\usage{
FINDATA
}
\description{
Data for estimating a VAR-process for Switzerland, necessary for optimizing over different asset classes
}
\keyword{datasets}
