% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpCF.R
\name{fpCF}
\alias{fpCF}
\title{First Pillar Cash Flows}
\usage{
fpCF(ret_age = 65, c_age, li, lg, s1, ret)
}
\arguments{
\item{ret_age}{optional, retirement age, can be set anywhere between 60 and 70 (default: 65)}

\item{c_age}{the investor's current age (assuming birthday is calculation-day)}

\item{li}{gross labor income at time 0 (in the last year before birthday)}

\item{lg}{labor growth rate (in real terms, constant)}

\item{s1}{vector consisting of two components: c(number of contribution years at age=c_age,historical average yearly income until c_age)}

\item{ret}{investment return scenarios (nominal)}
}
\value{
Vector of Cashflows as of pension starting age (ret_age) until age=122 (dim=c(122-retage,# scenarios))
}
\description{
Cash Flows are generated for the payment phase of the first pillar, starting in retirement (ret_age) until 122 because that is when all our
ficticious persons have died. We use "Russian" inflation adjustment for starting pension: average income is computed mixing nominal past and
real future income, brackets in pension table are not adjusted ==> running pension in real terms decreases by half the inflation rate
}
\examples{
data(ret)
fpCF(ret_age=65,c_age=42,li=100000,lg=0.01,s1=c(15,80000),ret[,,1:10])

}
