% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimalLC.R
\name{optimalLC}
\alias{optimalLC}
\title{Optimizer to maximize Expected Utility}
\usage{
optimalLC(
  initial_values,
  upper_bounds,
  lower_bounds,
  ret_age,
  ra,
  delta,
  beta,
  c_age,
  gender,
  gender_mortalityTable,
  w0,
  CF,
  li,
  lg,
  c1,
  s1,
  s2,
  s3,
  w2,
  rho2,
  rho3,
  ret,
  retr,
  psi,
  verbose = FALSE,
  warnings = FALSE,
  trace = 0,
  reltol = sqrt(.Machine$double.eps)
)
}
\arguments{
\item{initial_values}{Starting values c(ret_age, c, c2, nu2, nu3, alpha, w3)}

\item{upper_bounds}{Upper bounds for optimization}

\item{lower_bounds}{Lower bounds for optimization}

\item{...}{all other parameters}
}
\value{
Expected utility
}
\description{
Here we minimize negative expected utility (given by function `util()`). Parameters can be fixed by setting "tight" boundaries
}
\examples{
\dontrun{
data(ret);data(retr)
.load_parameters()
initial_values <- c(0.6, 0.12, 0.5, 0.2, 0.96, 0.25, 0.25, 0.25)
optimalLC(initial_values,upper_bounds=NULL,lower_bounds=NULL,ret_age=ret_age,ra=ra,delta=delta,beta=beta,c_age=c_age,gender=gender,gender_mortalityTable=gender_mortalityTable,
         w0=w0,CF=CF,li=li,lg=lg,c1=c1,s1=s1,s2=s2,s3=s3,w2=w2,rho2=rho2,rho3=rho3,ret=ret[,,1:10],retr=retr[,,1:10],psi=psi,trace=1,reltol=1e-5)

}
}
