% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ret}
\alias{ret}
\title{Asset returns (nominal) created internally}
\format{An 121x5x10000 array}
\source{
own calculation
}
\usage{
ret
}
\description{
10000 return matrices simulated by VARsim using V
}
\keyword{datasets}
