% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{retr}
\alias{retr}
\title{Asset returns (real) created internally}
\format{
An 121x5x10000 array
}
\source{
own calculation
}
\usage{
retr
}
\description{
correct returns by subtracting the inflation column.
}
\keyword{datasets}
