% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spCF.R
\name{spCF}
\alias{spCF}
\title{Second Pillar Cash Flows}
\usage{
spCF(
  ret_age = 65,
  nu2,
  c_age,
  c2,
  li,
  lg,
  w2 = setNames(c(0.3, 0.3, 0.3, 0.1, 0), c("msci", "b10", "recom", "libor", "infl")),
  ret,
  retr,
  s2,
  rho2,
  warnings = TRUE
)
}
\arguments{
\item{ret_age}{optional, retirement age, can be set anywhere between 60 and 70 (default: 65)}

\item{nu2}{fraction of second pillar savings that is converted to life-long pension}

\item{c_age}{the investor's current age (assuming birthday is calculation-day)}

\item{c2}{second pillar savings as fraction of gross income (still missing: health, a-fonds-perdu payments)}

\item{li}{gross labor income at time 0 (in the last year before birthday)}

\item{lg}{labor growth rate (in real terms, constant)}

\item{w2}{portfolio allocation (assumed to be fixed and not influenced by the decision maker)}

\item{ret}{investment return scenarios (nominal)}

\item{retr}{investment return scenarios (real)}

\item{s2}{savings in second pillar as of t=0}

\item{rho2}{conversion factor in second pillar for regular retirement age}

\item{warnings}{optional: should warnings be given? (default=TRUE)}
}
\value{
List with elements:
- a vector of lumpsum spendings (length = number of scenarios) and
- a matrix with annuity payments starting at ret_age until 122 (rows) with number of columns equal to the number of scenarios
- vector of wealth levels at retirement
}
\description{
Cash Flows are generated for the payment phase of the second pillar, starting at ret_age until 122 because that is when all our
ficticious persons have died. Contributions to the second pillar are doubled by the employer
}
\examples{
data(ret); data(retr)
sp_ex <- spCF(ret_age=65,nu2=.5,c_age=42,c2=.12,
           li=100000,lg=0.01,
           ret=ret[,,1:10],retr=retr[,,1:10],s2=300000,rho2=0.05)
sp_ex2 <- spCF(ret_age=65,nu2=0.01,c_age=64,c2=0.01,li=0,lg=0.01,
w2=setNames(c(1,0,0,0,0),c("msci","b10","recom","libor","infl")),
           ret=ret[,,1:10],retr=retr[,,1:10],s2=0,rho2=0.0000001)

}
