% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SPFret}
\alias{SPFret}
\title{Portfolio returns in second pillar created internally
for a pre-fixed average asset allocation}
\format{
An (69-18+1)x(70-60+1)x((70-20)+c_age)x10000 array
}
\source{
own calculation
}
\usage{
SPFret
}
\description{
A list (c_age: 18-69) of lists (ret_age: 60-70) that each contains a
matrix detailing the returns of the second pillar investments
in all 10000 scenarios
}
\keyword{datasets}
