% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxCFlumpsum.R
\name{taxCFlumpsum}
\alias{taxCFlumpsum}
\title{Calculate lumpsum after taxes}
\usage{
taxCFlumpsum(lumpsum, ret_age = 65, gender)
}
\arguments{
\item{lumpsum}{capital taken out at retirement from second pillar (vector?)}

\item{ret_age}{optional, retirement age, can be set anywhere between 60 and 70 (default: 65)}

\item{gender}{gender, 0=male and 1=female}
}
\value{

}
\description{
This function is supposed to subtract taxes on lumpsum payments from the second pillar at pension start
}
\examples{
taxCFlumpsum(lumpsum=1000000,ret_age = c(60,64,65),gender=0)
taxCFlumpsum(lumpsum=1000000,ret_age = c(60,64,65),gender=c(0,1))
taxCFlumpsum(lumpsum=c(0,210,300),ret_age = 64,gender=0)

}
