% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxCFret.R
\name{taxCFret}
\alias{taxCFret}
\title{Calculate total taxes for pension cash-flows}
\usage{
taxCFret(fpcf, totalcf, wealth)
}
\arguments{
\item{fpcf}{first pillar pension cash-flow}

\item{totalcf}{total pension cash-flow (including fpcf)}

\item{wealth}{wealth at 1.1. of tax year (can be negative)}
}
\value{
pension income after tax (vector or scalar)
}
\description{
Calculate total taxes for pension cash-flows
}
\examples{
taxCFret(fpcf=15000,totalcf=25000,wealth=100000)
taxCFret(fpcf=c(15000,15000),totalcf=c(30000,30000),wealth=c(100000,-120000))

}
