% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxCFwork.R
\name{taxCFwork}
\alias{taxCFwork}
\title{Calculate Income after taxes during savings phase}
\usage{
taxCFwork(income, wealth)
}
\arguments{
\item{income}{income bfore tax (vector or scalar)}

\item{wealth}{wealth at 1.1. of tax year}
}
\value{
income after tax (vector or scalar)
}
\description{
The current function calculates income after deduction taxes during the cash-flow phase and is an important
input parameter for the third pillar function (tpCFwork), where CF after tax is neded to determine consumption/saving
}
\examples{
taxCFwork(income=100000,wealth=c(300000,310000))

}
