% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpCFwork.R
\name{tpCFwork}
\alias{tpCFwork}
\title{Third Pillar Cash Flows during savings phase}
\usage{
tpCFwork(
  ret_age = 65,
  c_age,
  w3,
  free_cf_before_tax,
  retr,
  s3,
  w0,
  psi = 0.015,
  c,
  warnings = TRUE
)
}
\arguments{
\item{ret_age}{optional, retirement age, can be set anywhere between 60 and 70 (default: 65)}

\item{c_age}{the investor's current age (assuming birthday is calculation-day)}

\item{w3}{third pillar portfolio allocation (given either as vector (constant investment or as matrix with
entries) for all remaining years}

\item{free_cf_before_tax}{remaining cash flow after contributing to first and second pillar}

\item{retr}{investment return scenarios (real)}

\item{s3}{liquid wealth - invested in the third pillar (current assumption: no tax advantage for third pillar)}

\item{w0}{time c_age wealth that is not disposable, assumption: still available at retirement (no growth or decline), alternatively: expected wealth (that is not disposable) at retirement}

\item{psi}{optional, spread to take a loan/leverage for third pillar savings}

\item{c}{fraction of income that is consumed while still working (current assumption: constant)}

\item{warnings}{optional: should warnings be given? (default=TRUE)}
}
\value{
list with two elemnts:
-  consumption during savings years (matrix with dim=c(sav_years,# of Scenarios))
-  Development of wealth during savings years (matrix with dim=c(sav_years,# of Scenarios))
}
\description{
Cash Flows are generated for the savings phase of the third pillar, starting at c_age until ret_age
}
\examples{
data(ret); data(retr)
free_cf_before_tax <- c(81000.00,81810.00,82628.10,83454.38,84288.92,85131.81,85983.13,86842.96,87711.39,88588.51,
89474.39,90369.14,91272.83,92185.56,93107.41,94038.49,94978.87,95928.66,96887.95,97856.82,98835.39,99823.75,100821.98)
tpwork_ex <- tpCFwork(ret_age=65,c_age=42,w3=setNames(c(.25,.25,.25,.25,0),c("msci","b10","recom","libor","infl")),
           free_cf_before_tax=free_cf_before_tax,retr=retr[,,1:10],s3=300000,w0=300000,psi=0.015,c=0.6)
tpwork_ex2 <- tpCFwork(ret_age=65,c_age=42,w3=setNames(c(.30,.30,.30,.10,0),c("msci","b10","recom","libor","infl")),
           free_cf_before_tax=free_cf_before_tax,retr=retr[,,1:10],s3=0,w0=0,psi=0.015,c=1)

}
