% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{util}
\alias{util}
\title{Calculate utility of total consumption and bequest}
\usage{
util(ret_age, tw3, c, c2, nu2, nu3, ra, delta, alpha, beta, c_age, gender,
  gender_mortalityTable, w0, CF, li, lg, c1, s1, s2, s3, w2, rho2, rho3,
  ret, retr, psi, verbose = FALSE)
}
\arguments{
\item{ret_age}{Decision Variable: retirement age, can be set anywhere between 60 and 70 (default: 65)}

\item{tw3}{Decision Variable: third pillar portfolio allocation (given either as vector or as matrix with entries) for all years.
HERE: Choose only allocation to stocks, bonds and real estate, cash will be determined as fraction missing to sum up to one}

\item{c}{Decision Variable: fraction of income that is consumed while still working (current assumption: constant)}

\item{c2}{Decision Variable: second pillar savings as fraction of gross income (still missing: health, a-fonds-perdu payments)}

\item{nu2}{Decision Variable: fraction of second pillar savings that is converted to life-long pension}

\item{nu3}{Decision Variable: fraction of third pillar savings that is converted to life-long pension}

\item{ra}{Given variable: Risk Aversion of Agent}

\item{delta}{Given Variable: Time Preference}

\item{alpha}{Decision variable: Parameter to choose fraction of wealth NOT consumed during retirement - see file 'Consumption_3p.ods'}

\item{beta}{Given variable: Relative Weight of bequest utility}

\item{c_age}{Given variable: the investor's current age (assuming birthday is calculation-day)}

\item{gender}{Given variable: gender, 0=male and 1=female}

\item{gender_mortalityTable}{Given variable: MortalityTable to use for gender 0=male and 1=female, e.g. `baseTable(AVOe2005R.male)`}

\item{w0}{Given variable: time c_age wealth that is not disposable, assumption: still available at retirement (no growth or decline),
alternatively: expected wealth (that is not disposable) at retirement, stays the same over time}

\item{CF}{Given Variables: income shocks, such as inheritance (not currrently imlemented)}

\item{li}{Given variable: gross labor income at time 0 (in the last year before birthday)}

\item{lg}{Given variable: labor growth rate (in real terms, constant)}

\item{c1}{Given variable: first pillar savings as fraction of gross income}

\item{s1}{Given variable: vector consisting of two components: c(number of contribution years at age=c_age,historical average yearly income until c_age)}

\item{s2}{Given variable: savings in second pillar as of t=0}

\item{s3}{Given variable: liquid wealth - invested in the third pillar (current assumption: no tax advantage for third pillar)}

\item{w2}{Given variable: portfolio allocation in second pillar (assumed to be fixed and not influenced by the decision maker)}

\item{rho2}{Given variable: conversion factor in second pillar for regular retirement age}

\item{rho3}{Given variable: conversion factor in third pillar for regular retirement age}

\item{ret}{Given variable: investment return scenarios (nominal)}

\item{retr}{Given variable: investment return scenarios (real)}

\item{psi}{Given variable: spread to take a loan/leverage for third pillar savings}

\item{verbose}{optional: show additional information while calculating utility (default: FALSE)}
}
\value{
Expected utility
}
\description{
Description: Utility for total cash-flows
}
\examples{
data(ret); data(retr)
MortalityTables::mortalityTables.load("Austria_Annuities")

util_ex <- util(ret_age=65,c_age=42,
                tw3=setNames(c(.25,.25,.25),c("msci","b10","recom")),
                c=0.6,c2=.12,nu2=.5,nu3=0.01,ra=4,delta=0.02,alpha=0.96,beta=0.75,gender=0,
                gender_mortalityTable=MortalityTables::baseTable(AVOe2005R.male),
                w0=300000,CF=NULL,li=100000,lg=0.01,c1=0.07,s1=c(15,80000),s2=300000,s3=300000,
                w2=setNames(c(.30,.30,.30,.10,0),c("msci","b10","recom","libor","infl")),
                rho2=0.05,rho3=0.04,ret=ret[,,1:10],retr=retr[,,1:10],psi=0.015)
util_ex2 <- util(ret_age=65,c_age=64,
                tw3=setNames(c(1,0,0),c("msci","b10","recom")),
                c=1,c2=0,nu2=0,nu3=0,ra=0,delta=0,alpha=1,beta=0,gender=0,
                gender_mortalityTable=MortalityTables::baseTable(AVOe2005R.male),
                w0=0,CF=NULL,li=0,lg=0,c1=0,s1=c(0,0),s2=0,s3=0,
                w2=setNames(c(0,0,0,1,0),c("msci","b10","recom","libor","infl")),
                rho2=0.0001,rho3=0.0001,ret=ret[,,1:10],retr=retr[,,1:10],psi=0)

}
