% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{V}
\alias{V}
\title{VAR object for FINDATA}
\format{
An object (list) with many elements
}
\source{
own calculation
}
\usage{
V
}
\description{
The output of a VAR estimation using `lineVAR` from `tsDyn`.
}
\keyword{datasets}
