% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VARsim.R
\name{VARsim}
\alias{VARsim}
\title{Simulate from given (pre-estimated) VAR process}
\usage{
VARsim(
  V,
  simN = 1000,
  age.max = 122,
  age = 0,
  frequency = 4,
  s = NULL,
  covres = NULL
)
}
\arguments{
\item{V}{Output of a VAR Estimation (in package tsDyn)}

\item{simN}{optional, number of simulated return series to generate, defaults to 1000}

\item{age.max}{optional,maximum age, defaults to 122}

\item{age}{optional,starting age, defaults to one, as we usually want to generate all lifetime returns and select from them accordingly}

\item{frequency}{optional,defaults to 4 (quarterly), could also be 12 (monthly)}

\item{s}{optional, (from VAR.sim): Starting values (matrix of dimension lag x k) for the VAR to simulate, in our case this
defaults to the steady state calculated as $s=(diag(k)-B)^{-1} * c$, where k is the dimension of the VAR, B the coefficients
of parameters and equations and c the intercepts of the equations, all taken from V}

\item{covres}{optional, Variance-covariance matrix for the innovations, defaults to the covariance matrix of the VAR given by V as $var(V$residuals)/(V$T-2-V$k)*(V$T-2)$}
}
\value{
An array of dimension (T,k,simN) containing simN simulated return series for T years and k variables
}
\description{
Simulate from given (pre-estimated) VAR process
}
\examples{
data(V)
retPA <- VARsim(V,simN=10)

}
