% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimalLC.R
\name{optimalLC}
\alias{optimalLC}
\title{Optimiser to maximize Expected Utility}
\usage{
optimalLC(
  initial_values,
  upper_bounds,
  lower_bounds,
  ra,
  delta,
  beta,
  c_age,
  gender,
  gender_mortalityTable,
  w0,
  CF,
  li,
  lg,
  c1,
  s1,
  s2,
  s3,
  w2,
  rho2,
  rho3,
  ret,
  retr,
  psi,
  trace
)
}
\arguments{
\item{initial_values}{Starting values c(ret_age,c, c2, nu2, nu3, alpha, w3)}

\item{upper_bounds}{Upper bounds for optimization}

\item{lower_bounds}{Lower bounds for optimization}

\item{...}{all other parameters}
}
\value{
Expected utility
}
\description{
Here we minimize negative expected utility (given by function `util`). Parameters can be fixed by setting "tight" boundaries.
}
\examples{


}
