% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimalwc.R
\name{optimalwc}
\alias{optimalwc}
\title{Optimiser to maximize Expected Utility}
\usage{
optimalwc(
  initial_values,
  upper_bounds,
  lower_bounds,
  ret_age,
  c2,
  nu2,
  nu3,
  ra,
  delta,
  beta,
  c_age,
  gender,
  gender_mortalityTable,
  w0,
  CF,
  li,
  lg,
  c1,
  s1,
  s2,
  s3,
  w2,
  rho2,
  rho3,
  ret,
  retr,
  psi,
  verbose = FALSE,
  trace = 0
)
}
\arguments{
\item{initial_values}{Starting values c(c, alpha, w3)}

\item{upper_bounds}{Upper bounds for optimization}

\item{lower_bounds}{Lower bounds for optimization}

\item{...}{all other parameters}
}
\value{
Expected utility
}
\description{
Here we minimize negative expected utility (given by function `util`). Parameters can be fixed by setting "tight" boundaries.
}
\examples{

data(ret);data(retr)
.load_parameters()
initial_values <- c(0.5, 0.95, 0.25, 0.25, 0.25)
upper_bounds <- c(1, 1, 2, 2, 2)
lower_bounds <- c(0, 0, -2, -2, -2)
outwc <- optimalwc(initial_values,upper_bounds,lower_bounds,ret_age=ret_age,c2=c2,nu2=nu2,nu3=nu3,
         ra=18,delta=delta,beta=beta,c_age=c_age,gender=gender,gender_mortalityTable=gender_mortalityTable,
         w0=w0,CF=CF,li=li,lg=lg,c1=c1,s1=s1,s2=s2,s3=s3,w2=w2,rho2=rho2,rho3=rho3,ret=ret[,,1:10],retr=retr[,,1:10],psi=psi,trace=0)

}
