% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxCFwork.R
\name{taxCFwork}
\alias{taxCFwork}
\title{Calculate Income after taxes during savings phase}
\usage{
taxCFwork(income, liquid_wealth, illiquid_wealth)
}
\arguments{
\item{income}{income before tax (vector or scalar)}

\item{liquid_wealth}{liquid wealth at 1.1. of tax year (S3)}

\item{illiquid_wealth}{illiquid wealth at 1.1. of tax year (w0)}
}
\value{
net (after tax) income payments, specifying which part should be paid from cash-flows (income) and from liquid wealth
}
\description{
This function calculates income after deduction taxes during the cash-flow phase and is an important input parameter
for the third pillar function `tpCFwork()`, where CF after tax is needed to determine consumption/saving. It further
specifies which part of taxes should be paid from income and which part from liquid wealth
}
\examples{
taxCFwork(income=c(100000,100000,100000),liquid_wealth=c(100000,50000,-100000),illiquid_wealth=c(100000,150000,100000))

}
